% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonredundant.R
\name{ids_with_multiple_obs}
\alias{ids_with_multiple_obs}
\title{\code{ids_with_multiple_obs} returns a vector of values that occur multiple times}
\usage{
ids_with_multiple_obs(id)
}
\arguments{
\item{id}{A vector of dummy indicators}
}
\value{
A vector of repeated values
}
\description{
\code{ids_with_multiple_obs} returns a vector of values that occur multiple times
}
