% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonredundant.R
\name{nonredundant}
\alias{nonredundant}
\title{Nonredundant}
\usage{
nonredundant(iid, tid, w)
}
\arguments{
\item{iid}{A vector of group dummy indicators}

\item{tid}{A vector of time dummy indicators}

\item{w}{A vector of non-negative weights}
}
\value{
A list will be returned with the following named values:
        flag - Are there redundant dummy levels?
        nr - a listing of
}
\description{
\code{nonredundant} flags for redundant dummy levels
}
