% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress1.R
\name{regress1}
\alias{regress1}
\title{Regression}
\usage{
regress1(y, X)
}
\arguments{
\item{y}{The dependant variable}

\item{X}{A matrix or vector of independant variable(s)}
}
\value{
A list which contains X'X, the returned coefficients beta, and residuals res
}
\description{
\code{regress1} performs an OLS regression based on the projected variables y and X.
}
\examples{
hhid <- c("a","b","c")
tid <- c("1","2","3")
w <- c(1,1,1)
x1 <- c(1,2,3)
y <- c(2,4,6)
struc <- projdummies(hhid, tid, w)
x1_projected <- projvar(x1, struc)
y_projected <- projvar(y, struc)
reg <- regress1(y_projected, x1_projected)
}
