% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonredundant.R
\name{where_id_with_single_obs}
\alias{where_id_with_single_obs}
\title{\code{where_id_with_single_obs} returns a true/false vector, 
Does this dummy occur multiple times?}
\usage{
where_id_with_single_obs(id)
}
\arguments{
\item{id}{A vector of dummy indicators}
}
\value{
A true/false vector
}
\description{
\code{where_id_with_single_obs} returns a true/false vector, 
Does this dummy occur multiple times?
}
