% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.R
\name{avar}
\alias{avar}
\title{Asymptotic variance of Estimator}
\usage{
avar(X, e, group = NULL, J = NULL)
}
\arguments{
\item{X}{A matrix or vector of independant variable(s)}

\item{e}{The residuals from the regression}

\item{group}{(optional) The cluster identifier (hhid from \code{projdummies})}

\item{J}{(optional) This is assumed to be X'X, and can be input if pre-calculated}
}
\value{
A matrix of the covariates
}
\description{
\code{avar} calculated the asymptotic variance of the regression estimation
}
\examples{
hhid <- c("a","b","c")
tid <- c("1","2","3")
w <- c(1,1,1)
x1 <- c(1,2,3)
y <- c(2,4,6)
struc <- projdummies(hhid, tid, w)
x1_projected <- projvar(x1, struc)
y_projected <- projvar(y, struc)
reg <- regress1(y_projected, x1_projected)
matCov <- avar(X, reg$res, struc$hhid, reg$XX)
}
